requirements.txt
flask
python-dotenv
pandas
numpy